requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
pip==20.2.4
PyYAML==5.3.1
requests==2.25.0
setuptools==50.3.2
urllib3==1.26.2
wheel==0.35.1
